Shadow MMR theGraph (#429)
* Move thegraph-liketh to assistants

* Delete redundant files

* Add merkle-mountain-range for shadow component

* Update lock

* Bump thegraph-liketh and gql_client

* Rename merkle-mountain-range to mmr-client

* Add shadow-liketh

* Test shadow mmr thegraph

* Add mmr position query graphql

* Try query mmr root

* gen mmr proof

* Update shadow parcel

* Completed shadow update

* Fix compile

* Merge master and fix conflicts

* Update deps

* Change naming and deps. Pick mmr functions from shadow

* Fix fixed array deserialize

* clippy

* Fix compile

* Clippy

* Fix wrong mmr

* Merge new subxt

* Fix call

* Merge master

* Add common toolkit crate and correct enpoint function

* Bump deps

* Update for bridge darwinia-ethereum

* Format log

* Format log

* Format log

* Test bridge

* Test gen mmr proof

* Test mmr root

* add mmr client (#457)

* Test shadow

* Lint

* Lint

* Lint

* Remove shadow component

* Better naming for shadow and use support-mmr

* Update to support new shadow https://github.com/darwinia-network/shadow/pull/166

* Remove useless extern

* Update shadow types

* Fix shadow types

* Update deps

* Update deps

* use support toolkit

* Bump deps

* Try fix redeem

* Fix shadow issue and test pangolin-ropsten

* Test darwinia-ethereum

* Fix wrong features

* Format log

* Waiting some blocks for affirm and fix redeem order error

* Fix wrong break condition and rename fortx to for_tx

* Try fix ci

* Clearer log

* Fix bug. when the queried block is smaller than the data stored in thegraph, it will crash

* Set safe number to 12 for ethereum

* Fix affirm bug

* Better than better

* Correct last leaf and fix merkle proof order

* Use last confirmed

* Test verify mmr proof

* Fix wrong node order

* Remove sorter

* Fix some transactions in same block bug for redeem and increase limit for affirm

* Fix check service stalled behavior

* Rename tx_hash to block_hash

* Fix subxt codegen bug for EthereumBacking::VerifiedProof storage

* Trigger CI

* Update config template

* Trigger CI

* Trigger CI

* Fix pangolin subql deploy issues and unify all subql configurations

* unify all subql configurations

* Update follow suggestions


* Update follow suggestions

+++ bridges/darwinia-ethereum/bridge/src/command/types/affirm.rs
@@ -23,7 +23,7 @@ pub enum AffirmOpts {
         mode: AffirmMode,
         /// The block number for ethereum
         #[structopt(long)]
-        block: Option<u64>,
+        block: Option<u32>,
         /// Raw parcel json data
         #[structopt(long = "raw")]
         raw_json: Option<String>,
